Abstract pageRenderer from server.jsx
