+++ app/config/app.js
@@ -1,0 +1,6 @@
+export default {
+  host: process.env.HOSTNAME || 'localhost',
+  port: process.env.PORT || '3000',
+  trackingID: null // Replace with 'UA-########-#' or similar to enable tracking
+};
+
