+++ app/pages/index.js
@@ -1,3 +1,5 @@
 export { default as Vote} from 'pages/Vote';
 export { default as Dashboard } from 'pages/Dashboard';
+export { default as LoginOrRegister } from 'pages/LoginOrRegister';
+export { default as About } from 'pages/About';
