+++ app/server.jsx
@@ -1,49 +1,17 @@
 import axios from 'axios';
 import React from 'react';
-import { renderToString } from 'react-dom/server';
-import { createMemoryHistory, match, RouterContext } from 'react-router';
-import { Provider } from 'react-redux';
+import { createMemoryHistory, match } from 'react-router';
 import createRoutes from 'routes';
 import configureStore from 'store/configureStore';
 import * as types from 'types';
 import preRenderMiddleware from 'middlewares/preRenderMiddleware';
-import header from 'components/Meta';
+import config from 'config/app';
+import pageRenderer from 'utils/pageRenderer';
 
-const clientConfig = {
-  host: process.env.HOSTNAME || 'localhost',
-  port: process.env.PORT || '3000'
-};
+
 
 // configure baseURL for axios requests (for serverside API calls)
-axios.defaults.baseURL = `http://${clientConfig.host}:${clientConfig.port}`;
-
-
-const analtyicsScript =
-  typeof trackingID === "undefined" ? ``
-  :
-  `<script>
-    (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
-    (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
-    m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
-    })(window,document,'script','https://www.google-analytics.com/analytics.js','ga');
-    ga('create', ${trackingID}, 'auto');
-    ga('send', 'pageview');
-  </script>`;
-
-
-/*
- * To Enable Google analytics simply replace the hashes with your tracking ID
- * and move the constant to above the analtyicsScript constant.
- *
- * Currently because the ID is declared beneath where is is being used, the
- * declaration will get hoisted to the top of the file.
- * however the assignement  does not, so it is undefined for the type check above.
- * https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Statements/var#var_hoisting
- */
-const trackingID  = "'UA-########-#'";
-
-
-
+axios.defaults.baseURL = `http://${config.host}:${config.port}`;
 
 /*
  * Export render function to be used in server/config/routes.js
@@ -96,30 +64,8 @@ export default function render(req, res) {
       preRenderMiddleware(props)
       .then(data => {
         store.dispatch({ type: types.REQUEST_SUCCESS, data });
-        const componentHTML = renderToString(
-          <Provider store={store}>
-            <RouterContext {...props} />
-          </Provider>
-        );
-
-        const initialState = store.getState();
-
-        res.status(200).send(`
-          <!doctype html>
-          <html ${header.htmlAttributes.toString()}>
-            <head>
-              ${header.title.toString()}
-              ${header.meta.toString()}
-              ${header.link.toString()}
-            </head>
-            <body>
-              <div id="app">${componentHTML}</div>
-              <script>window.__INITIAL_STATE__ = ${JSON.stringify(initialState)};</script>
-              ${analtyicsScript}
-              <script type="text/javascript" charset="utf-8" src="/assets/app.js"></script>
-            </body>
-          </html>
-        `);
+        const html = pageRenderer(store, props);
+        res.status(200).send(html);
       })
       .catch(err => {
         console.error(err);
